requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
line-bot-sdk==2.4.2
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
yarl==1.9.2